feat: Implement TAV state context
This commit implements the plan to fix the "Chat with TAV" button in the footer. It introduces a React Context Provider for managing the TAV state, ensuring state is shared across components. This includes creating the TavaraStateContext, updating the useTavaraState hook to consume the context, adding the TavaraStateProvider to AppProviders, and testing the functionality.
