+++ src/components/providers/AppProviders.tsx
@@ -3,6 +3,7 @@ import { Toaster as Sonner } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { AuthProvider } from "@/components/providers/AuthProvider";
+import { TavaraStateProvider } from "@/components/tav/hooks/TavaraStateContext";
 import { BrowserRouter } from "react-router-dom";
 
 const queryClient = new QueryClient({
@@ -23,7 +24,9 @@ export function AppProviders({ children }: { children: React.ReactNode }) {
         <Sonner />
         <BrowserRouter>
           <AuthProvider>
-            {children}
+            <TavaraStateProvider>
+              {children}
+            </TavaraStateProvider>
           </AuthProvider>
         </BrowserRouter>
       </TooltipProvider>
