+++ src/components/tav/index.ts
@@ -4,6 +4,7 @@ export { RoleBasedContent } from './RoleBasedContent';
 export { ManualNudgeService } from './ManualNudgeService';
 export { assistantSupabase } from './assistantSupabase';
 export { useTavaraState } from './hooks/useTavaraState';
+export { TavaraStateProvider } from './hooks/TavaraStateContext';
 export { useReturningUser } from './hooks/useReturningUser';
 export { useFormDetection } from './hooks/useFormDetection';
 export { useConversationalForm } from './hooks/useConversationalForm';
